requirements.txt
alpaca-py==0.20.3
annotated-types==0.6.0
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
msgpack==1.0.8
numpy==1.26.4
pandas==2.2.2
pydantic==2.7.0
pydantic_core==2.18.1
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
six==1.16.0
sseclient-py==1.8.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
websockets==12.0
backtrader
Django==5.0.6
python-dotenv
celery
redis
psycopg2
django-unfold
TA-Lib
django-webpack-loader
djangorestframework
yfinance
pyyaml
pytickersymbols